extract fuzzing strategies tests to separate test file
