+++ tests/integration/fuzzing_strategies/invalid_calls_test.cairo
