+++ tests/integration/fuzzing_strategies/multiple_learning_steps_test.cairo
